requirements.txt

django==4.2
django-rest-framework
djangorestframework-simplejwt
drf-spectacular

psycopg2-binary
django-celery
django-celery-beat

django-redis

redis
celery

#jwt
